implemented Builders for floor and building. modified List to Map as really the fields are accessed more by their values. may actually change the field to remove fieldname later on i guess. not really needed now that it is map

+++ FireFacilAuto/src/main/java/com/FireFacilAuto/domain/entity/floors/PossibleFloorFields.java
@@ -11,8 +11,8 @@ public class PossibleFloorFields {
         fieldMap.put("floorNo", Integer.class);
         fieldMap.put("isUnderGround", Boolean.class);
         fieldMap.put("floorClassification", Integer.class);
-        fieldMap.put("floorSpecification", Double.class);
-        fieldMap.put("floorArea",Integer.class);
+        fieldMap.put("floorSpecification", Integer.class);
+        fieldMap.put("floorArea",Double.class);
         fieldMap.put("floorMaterial",Integer.class);
         fieldMap.put("floorWindowAvailability",Boolean.class);
     }
